+++ src/tasks/atom.rs
@@ -55,7 +55,7 @@ pub fn sync() {
 
     for ext in config.install {
         if !pkgs.contains(&ext) {
-            utils::process::command_spawn_wait(COMMAND, &["install", "--compatible", "--production", " --quiet", &ext])
+            utils::process::command_spawn_wait(COMMAND, &["install", "--compatible", "--production", "--quiet", &ext])
                 .expect(ERROR_MSG);
         }
     }