+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* pkg: atom: fix `apm install` bug caused by accidental whitespace
+
 ## [0.4.0] - 2018-03-03
 
 ### Added