+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.4.1] - 2018-03-04
+
 ### Fixed
 
 * pkg: atom: fix `apm install` bug caused by accidental whitespace